Support an instance association with session records
An instance is a unique combination of one or more of the following
instance-specific values: port, protocol, or name.

This allows a single database to track records for multiple instances
with the ability to query on them later.

Allowing multiple methods to identify the instance should suit a variety
of use-cases, including running multiple instances on a single system or
collecting records from multiple external systems.