Add IPv6 columns to track future VPN IPv6 client assignment
This column is currently unused as OpenVPN does not expose client-issued
v6 details in the env-vars available to scripts. When this issue is
fixed (or with a patched OpenVPN version) this project will be updated
to populate this column.

By adding the column now, it avoids the need for a schema update to
support this new feature.